requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
dateparser==0.7
djangorestframework==3.12.2
idna==2.10
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
tqdm==4.46.0
urllib3==1.25.9
wincertstore==0.2
pynamodb==4.3.3
PyJWT==1.7.1
drf-spectacular==0.12.0
